chest_xray_medical/requirements.txt
torch==2.1.0
torchvision==0.16.0
gradio==4.16.0
opencv-python==4.8.1.78
numpy==1.24.3
matplotlib==3.8.2
Pillow==10.0.0
albumentations==1.3.1
ultralytics==8.0.198
scikit-learn==1.3.2
huggingface-hub==0.19.0
torchaudio==2.1.0